fix: re-write using cross-compat render function
Using a SFC lead us to unwanted bundle results when building the library. Apparently the template compiler used was the one for vue3. Manually changing rollup plugin vue compiler option to explicitly use other vue-template-compiler version did not change the results so we might be missing something.

The alternative implemented, was to convert the component into a pure render function, and middleware it with vue-demi isVue3 for conditionally assigning the data objects. This way there's not transformation when bundling, only minification and babel and the exported component is able to run everywhere as expected.